+++ cmd/monitor/streaming/main.go
@@ -19,11 +19,12 @@ import (
 	"github.com/erda-project/erda/pkg/common"
 
 	// modules
+	_ "github.com/erda-project/erda/modules/core/monitor/log/storage"
+	_ "github.com/erda-project/erda/modules/core/monitor/log/storage/v2"
 	_ "github.com/erda-project/erda/modules/core/monitor/metric/storage"
 	_ "github.com/erda-project/erda/modules/monitor/alert/storage/alert-record"
 	_ "github.com/erda-project/erda/modules/monitor/apm/browser"
-	_ "github.com/erda-project/erda/modules/monitor/core/logs/storage"
-	_ "github.com/erda-project/erda/modules/monitor/core/logs/storage/v2"
+	_ "github.com/erda-project/erda/modules/monitor/notify/storage/notify-record"
 	_ "github.com/erda-project/erda/modules/monitor/trace/storage"
 
 	// providers
@@ -33,10 +34,6 @@ import (
 	_ "github.com/erda-project/erda-infra/providers/kafka"
 	_ "github.com/erda-project/erda-infra/providers/mysql"
 	_ "github.com/erda-project/erda-infra/providers/pprof"
-
-	// storage record
-	_ "github.com/erda-project/erda/modules/monitor/alert/storage/alert-record"
-	_ "github.com/erda-project/erda/modules/monitor/notify/storage/notify-record"
 )
 
 func main() {
