+++ modules/core/monitor/log/query/query.go
@@ -22,12 +22,12 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/erda-project/erda-proto-go/core/monitor/log/query/pb"
 	"github.com/pkg/errors"
 	"github.com/scylladb/gocqlx"
 	"github.com/scylladb/gocqlx/qb"
 
-	"github.com/erda-project/erda/modules/monitor/core/logs/schema"
+	"github.com/erda-project/erda-proto-go/core/monitor/log/query/pb"
+	"github.com/erda-project/erda/modules/core/monitor/log/schema"
 )
 
 var ErrEmptyLogMeta = errors.New("empty log meta record")
