+++ modules/core/monitor/log/schema/cql.go
