+++ modules/core/monitor/log/storage/provider.go
@@ -19,7 +19,6 @@ import (
 	"time"
 
 	"github.com/bluele/gcache"
-	"github.com/sirupsen/logrus"
 
 	"github.com/erda-project/erda-infra/base/logs"
 	"github.com/erda-project/erda-infra/base/servicehub"
@@ -28,10 +27,10 @@ import (
 	mutex "github.com/erda-project/erda-infra/providers/etcd-mutex"
 	"github.com/erda-project/erda-infra/providers/kafka"
 	"github.com/erda-project/erda-infra/providers/mysql"
-	"github.com/erda-project/erda/modules/monitor/core/logs/schema"
+	"github.com/erda-project/erda/modules/core/monitor/log/schema"
 )
 
-const selector = "logs-store-v2"
+const selector = "log-store"
 
 type define struct{}
 
@@ -74,7 +73,7 @@ type provider struct {
 	EtcdMutexInf mutex.Interface
 	output       writer.Writer
 	ttl          ttlStore
-	schema       *schema.CassandraSchema
+	schema       schema.LogSchema
 	cache        gcache.Cache
 }
 
@@ -94,7 +93,7 @@ func (p *provider) Init(ctx servicehub.Context) error {
 		Log:           p.Log.Sub("ttlStore"),
 	}
 
-	p.schema, err = schema.NewCassandraSchema(cass, p.Log.Sub("logSchema"), schema.WithMutexKey(selector))
+	p.schema, err = schema.NewCassandraSchema(cass, p.Log.Sub("logSchema"))
 	if err != nil {
 		return err
 	}
@@ -105,15 +104,10 @@ func (p *provider) Init(ctx servicehub.Context) error {
 }
 
 func (p *provider) Run(ctx context.Context) error {
-	// create default
-	if err := p.schema.CreateDefault(); err != nil {
-		return fmt.Errorf("create default failed: %w", err)
-	}
-
 	go p.schema.RunDaemon(ctx, p.Cfg.Output.LogSchema.OrgRefreshInterval, p.EtcdMutexInf)
 	go p.ttl.Run(ctx, p.Cfg.Output.Cassandra.TTLReloadInterval)
 	go p.startStoreMetaCache(ctx)
-	if err := p.Kafka.NewConsumer(&p.Cfg.Input, p.invokeV2); err != nil {
+	if err := p.Kafka.NewConsumer(&p.Cfg.Input, p.invoke); err != nil {
 		return err
 	}
 
@@ -121,27 +115,6 @@ func (p *provider) Run(ctx context.Context) error {
 	case <-ctx.Done():
 	}
 	return nil
-}
-
-func (p *provider) startStoreMetaCache(ctx context.Context) {
-	ticker := time.NewTicker(p.Cfg.Output.Cassandra.CacheStoreInterval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ticker.C:
-			p.storeMetaCache()
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-func (p *provider) storeMetaCache() {
-	for _, meta := range p.cache.GetALL(true) {
-		if err := p.output.Write(meta); err != nil {
-			logrus.Errorf("fail to write log meta: %s", err)
-		}
-	}
 }
 
 func init() {
