+++ modules/core/monitor/log/storage/v2/monitoring.go
@@ -11,7 +11,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package storage
+package storagev2
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
