+++ modules/core/monitor/log/storage/v2/storage.go
@@ -11,7 +11,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package storage
+package storagev2
 
 import (
 	"bytes"
@@ -24,8 +24,9 @@ import (
 	"unsafe"
 
 	"github.com/erda-project/erda-infra/providers/cassandra"
-	"github.com/erda-project/erda/modules/monitor/core/logs"
-	"github.com/erda-project/erda/modules/monitor/core/logs/schema"
+	logmodule "github.com/erda-project/erda/modules/core/monitor/log"
+	"github.com/erda-project/erda/modules/core/monitor/log/pb"
+	"github.com/erda-project/erda/modules/core/monitor/log/schema"
 )
 
 func (p *provider) createLogStatementBuilder() cassandra.StatementBuilder {
@@ -43,16 +44,16 @@ type LogStatement struct {
 
 func (ls *LogStatement) GetStatement(data interface{}) (string, []interface{}, error) {
 	switch data.(type) {
-	case *logs.Log:
-		return ls.p.getLogStatement(data.(*logs.Log), ls.gzipWriter)
-	case *logs.LogMeta:
-		return ls.p.getMetaStatement(data.(*logs.LogMeta))
+	case *pb.Log:
+		return ls.p.getLogStatementV2(data.(*pb.Log), ls.gzipWriter)
+	case *logmodule.LogMeta:
+		return ls.p.getMetaStatement(data.(*logmodule.LogMeta))
 	default:
 		return "", nil, fmt.Errorf("value %#v must be Log or LogMeta", data)
 	}
 }
 
-func (p *provider) getLogStatement(log *logs.Log, reusedWriter *gzip.Writer) (string, []interface{}, error) {
+func (p *provider) getLogStatementV2(log *pb.Log, reusedWriter *gzip.Writer) (string, []interface{}, error) {
 	ttl := p.ttl.GetSecondByKey(log.Tags[diceOrgNameKey])
 
 	var requestID *string // request_id 字段不存在时为null，所以使用指针
@@ -73,7 +74,7 @@ func (p *provider) getLogStatement(log *logs.Log, reusedWriter *gzip.Writer) (st
 	cql := fmt.Sprintf(`INSERT INTO %s (source, id, stream, time_bucket, timestamp, offset, content, level, request_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?) USING TTL ?;`, table)
 	return cql, []interface{}{
 		log.Source,
-		log.ID,
+		log.Id,
 		log.Stream,
 		trncateDate(log.Timestamp),
 		log.Timestamp,
@@ -85,7 +86,7 @@ func (p *provider) getLogStatement(log *logs.Log, reusedWriter *gzip.Writer) (st
 	}, nil
 }
 
-func (p *provider) getMetaStatement(meta *logs.LogMeta) (string, []interface{}, error) {
+func (p *provider) getMetaStatement(meta *logmodule.LogMeta) (string, []interface{}, error) {
 	ttl := p.ttl.GetSecondByKey(meta.Tags[diceOrgNameKey])
 	cql := `INSERT INTO spot_prod.base_log_meta (source, id, tags) VALUES (?, ?, ?) USING TTL ?;`
 	return cql, []interface{}{
