+++ modules/extensions/loghub/index/query/log_query.go
@@ -24,7 +24,7 @@ import (
 	"github.com/recallsong/go-utils/encoding/jsonx"
 
 	"github.com/erda-project/erda-infra/providers/i18n"
-	"github.com/erda-project/erda/modules/monitor/core/logs"
+	logs "github.com/erda-project/erda/modules/core/monitor/log"
 )
 
 // Tag .
