+++ modules/extensions/loghub/index/query/log_query_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/recallsong/go-utils/encoding/jsonx"
 
-	"github.com/erda-project/erda/modules/monitor/core/logs"
+	logs "github.com/erda-project/erda/modules/core/monitor/log"
 )
 
 func Example_mergeLogSearch() {
